refactor: Dedicated module for duplicate requests test
